feat: angular v20 update
- updated angular to latest version (v20).
This included a cleanup to the angular workspace
to be similar to a newly generated workspace.
In the process the missing demo-dashboard application
was added and can now be used again.
- run angular migrations of control flow and standalone
