--- demo-dashboard/app/app.css
@@ -1,37 +1,0 @@
-.grid-item {
-	border: solid 1px #ccc;
-}
-
-.handle {
-	width: 100%;
-	height: 50px;
-	background: #eee;
-	color: black;
-	font-weight: 500;
-	padding: 15px;
-	box-sizing: border-box;
-}
-
-.btn {
-	margin-left: 10px;
-}
-
-.box-header-btns {
-    top: 15px;
-    right: 10px;
-    cursor: pointer;
-    position: absolute;
-}
-
-.navbar-default {
-	background: black;
-	border: 0px;
-}
-
-.navbar-default .navbar-brand {
-	color: white;
-}
-
-body {
-	background-color: #f5f2f2;
-}
