--- demo-dashboard/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "angular2-quickstart",
-  "scripts": {
-    "tsc": "tsc",
-    "tsc:w": "tsc -w",
-    "lite": "lite-server",
-    "start": "concurrent \"npm run tsc:w\" \"npm run lite\" "
-  },
-  "license": "ISC",
-  "dependencies": {
-    "@angular/common": "2.4.4",
-    "@angular/compiler": "2.4.4",
-    "@angular/core": "2.4.4",
-    "@angular/platform-browser": "2.4.4",
-    "@angular/platform-browser-dynamic": "2.4.4",
-    "angular2-grid": "^2.0.0",
-    "es6-shim": "^0.35.1",
-    "reflect-metadata": "^0.1.3",
-    "rxjs": "5.0.1",
-    "systemjs": "0.19.29",
-    "zone.js": "^0.7.2"
-  },
-  "devDependencies": {
-    "concurrently": "^2.1.0",
-    "lite-server": "^2.2.0",
-    "typescript": "^2.0.10"
-  }
-}
