--- demo-dashboard/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "system",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "removeComments": false,
-    "noImplicitAny": false,
-    "lib": [
-        "es2015",
-        "dom"
-    ]
-  },
-  "exclude": [
-    "node_modules"
-  ]
-}
