+++ package.json
@@ -45,32 +45,41 @@
   },
   "homepage": "https://github.com/BTMorton/angular2-grid",
   "private": true,
+  "prettier": {
+    "printWidth": 100,
+    "singleQuote": true,
+    "overrides": [
+      {
+        "files": "*.html",
+        "options": {
+          "parser": "angular"
+        }
+      }
+    ]
+  },
   "dependencies": {
-    "@angular/animations": "^17.1.1",
-    "@angular/common": "^17.1.1",
-    "@angular/compiler": "^17.1.1",
-    "@angular/core": "^17.1.1",
-    "@angular/forms": "^17.1.1",
-    "@angular/platform-browser": "^17.1.1",
-    "@angular/platform-browser-dynamic": "^17.1.1",
-    "@angular/router": "^17.1.1",
-    "bootstrap": "^3.3.7",
-    "rxjs": "~7.5.7",
-    "tslib": "^2.4.1",
-    "zone.js": "~0.14.3"
+    "@angular/common": "^20.3.0",
+    "@angular/compiler": "^20.3.0",
+    "@angular/core": "^20.3.0",
+    "@angular/forms": "^20.3.0",
+    "@angular/platform-browser": "^20.3.0",
+    "@angular/router": "^20.3.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.1",
-    "@angular/cli": "^17.1.1",
-    "@angular/compiler-cli": "^17.1.1",
-    "@types/jasmine": "~4.3.0",
-    "@types/node": "^18.11.9",
-    "jasmine-core": "~4.5.0",
-    "karma-chrome-launcher": "~3.1.1",
+    "@angular/build": "^20.3.3",
+    "@angular/cli": "^20.3.3",
+    "@angular/compiler-cli": "^20.3.0",
+    "@types/jasmine": "~5.1.0",
+    "jasmine-core": "~5.9.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "ng-packagr": "^17.1.2",
-    "typescript": "~5.3.3"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "ng-packagr": "^20.3.0",
+    "typescript": "~5.9.2"
   }
 }
