+++ projects/angular2-grid/README.md
@@ -1,24 +1,63 @@
-# AngularGrid
+# Angular2Grid
 
-This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 12.1.1.
+This project was generated using [Angular CLI](https://github.com/angular/angular-cli) version 20.3.0.
 
 ## Code scaffolding
 
-Run `ng generate component component-name --project angular2-grid` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project angular2-grid`.
-> Note: Don't forget to add `--project angular2-grid` or else it will be added to the default project in your `angular.json` file. 
+Angular CLI includes powerful code scaffolding tools. To generate a new component, run:
 
-## Build
+```bash
+ng generate component component-name
+```
 
-Run `ng build angular2-grid` to build the project. The build artifacts will be stored in the `dist/` directory.
+For a complete list of available schematics (such as `components`, `directives`, or `pipes`), run:
 
-## Publishing
+```bash
+ng generate --help
+```
 
-After building your library with `ng build angular2-grid`, go to the dist folder `cd dist/angular2-grid` and run `npm publish`.
+## Building
+
+To build the library, run:
+
+```bash
+ng build angular2-grid
+```
+
+This command will compile your project, and the build artifacts will be placed in the `dist/` directory.
+
+### Publishing the Library
+
+Once the project is built, you can publish your library by following these steps:
+
+1. Navigate to the `dist` directory:
+   ```bash
+   cd dist/angular2-grid
+   ```
+
+2. Run the `npm publish` command to publish your library to the npm registry:
+   ```bash
+   npm publish
+   ```
 
 ## Running unit tests
 
-Run `ng test angular2-grid` to execute the unit tests via [Karma](https://karma-runner.github.io).
+To execute unit tests with the [Karma](https://karma-runner.github.io) test runner, use the following command:
 
-## Further help
+```bash
+ng test
+```
 
-To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI Overview and Command Reference](https://angular.io/cli) page.
+## Running end-to-end tests
+
+For end-to-end (e2e) testing, run:
+
+```bash
+ng e2e
+```
+
+Angular CLI does not come with an end-to-end testing framework by default. You can choose one that suits your needs.
+
+## Additional Resources
+
+For more information on using the Angular CLI, including detailed command references, visit the [Angular CLI Overview and Command Reference](https://angular.dev/tools/cli) page.
