+++ projects/angular2-grid/package.json
@@ -1,6 +1,6 @@
 {
   "name": "angular2-grid",
-  "version": "11.0.3",
+  "version": "12.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/BTMorton/angular2-grid.git"
@@ -33,10 +33,11 @@
   },
   "homepage": "https://github.com/BTMorton/angular2-grid",
   "peerDependencies": {
-    "@angular/common": "^17.0.0",
-    "@angular/core": "^17.0.0"
+    "@angular/common": "^20.3.0",
+    "@angular/core": "^20.3.0"
   },
   "dependencies": {
-    "tslib": "^2.4.1"
-  }
+    "tslib": "^2.3.0"
+  },
+  "sideEffects": false
 }
