+++ projects/angular2-grid/src/lib/components/NgGridPlaceholder.ts
@@ -1,18 +1,19 @@
 import { NgGrid } from '../directives/NgGrid';
 import { NgGridItemPosition, NgGridItemSize } from '../interfaces/INgGrid';
-import { Component, ElementRef, Renderer2, OnInit } from '@angular/core';
+import { Component, ElementRef, Renderer2, OnInit, inject } from '@angular/core';
 
 @Component({
     selector: 'ng-grid-placeholder',
     template: ''
 })
 export class NgGridPlaceholder implements OnInit {
+    private _ngEl = inject(ElementRef);
+    private _renderer = inject(Renderer2);
+
     private _size: NgGridItemSize;
     private _position: NgGridItemPosition;
     private _ngGrid: NgGrid;
     private _cascadeMode: string;
-
-    constructor(private _ngEl: ElementRef, private _renderer: Renderer2) { }
 
     public registerGrid(ngGrid: NgGrid) {
         this._ngGrid = ngGrid;
