+++ projects/angular2-grid/src/lib/modules/NgGrid.module.ts
@@ -3,8 +3,11 @@ import { NgGrid } from '../directives/NgGrid';
 import { NgGridItem } from '../directives/NgGridItem';
 import { NgGridPlaceholder } from '../components/NgGridPlaceholder';
 
+/**
+ * @deprecated, please prefer importing the standalone components instead.
+ */
 @NgModule({
-    declarations: [NgGrid, NgGridItem, NgGridPlaceholder],
+    imports: [NgGrid, NgGridItem, NgGridPlaceholder],
     exports: [NgGrid, NgGridItem]
 })
 export class NgGridModule {}
