+++ src/app/app.ts
@@ -1,5 +1,6 @@
 import { Component, ViewChild, AfterViewInit } from '@angular/core';
-import { NgGridConfig, NgGridItemConfig, NgGridItemEvent, NgGrid } from 'angular2-grid';
+import { FormsModule } from '@angular/forms';
+import { NgGridConfig, NgGridItemConfig, NgGridItemEvent, NgGrid, NgGridItem } from 'angular2-grid';
 
 interface Box {
     id: number;
@@ -7,11 +8,12 @@ interface Box {
 }
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+    selector: 'app-root',
+    templateUrl: './app.html',
+    styleUrls: ['./app.css'],
+    imports: [FormsModule, NgGrid, NgGridItem]
 })
-export class AppComponent implements AfterViewInit {
+export class App implements AfterViewInit {
     @ViewChild(NgGrid)
     private grid: NgGrid;
     public curNum = 10;
